Add support for games that use a generic soundlatch device
Merged dataeast and taito systems to make use of the generic latch.
Games that are not inside the .dat but are eligible will try to load
a generic setting. Most games that use a z80 cpu for audio should work as is.

Modified tracklist load. Track lists are loaded automatically when
a file with the driver name is inside the tracklist folder.